Move tag helper function to their own helper module
